% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_archimed.R
\name{get_free_ram}
\alias{get_free_ram}
\title{Get available RAM}
\usage{
get_free_ram()
}
\value{
The amount of RAM available on the system in megabytes
}
\description{
Get the RAM available on the system
}
\details{
The function calls wmic for windows and awk for Linux. It is not
compatible with Mac OS yet.
}
\examples{
get_free_ram()
}
