% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_meteo.R
\name{import_meteo}
\alias{import_meteo}
\title{Import ARCHIMED meteo file}
\usage{
import_meteo(x)
}
\arguments{
\item{x}{The file path or connexion to the meteo file}
}
\value{
The meteo file in a more conveniant format for R usage
}
\description{
Import ARCHIMED meteo file
}
\details{
The function uses \code{\link[data.table:fread]{data.table::fread()}} under the hood, so the
\code{x} argument should match the input format as for this function.
}
\examples{
 \dontrun{
 import_meteo("output/meteo.csv")
 }

}
