% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{read_config}
\alias{read_config}
\title{Read configuration}
\usage{
read_config(file, parameter = NULL, format = c("list", "tibble"))
}
\arguments{
\item{file}{The path to the ARCHIMED configuration file}

\item{parameter}{A vector or list of parameter names. If \code{NULL}, returns all parameters.}

\item{format}{The output format. Should be either "list" or "tibble".}
}
\value{
The ARCHIMED configuration parameters as a list or a \code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Read the ARCHIMED configuration file parameters and
return one, several or all parameters as a list or as a \code{\link[tibble:tibble]{tibble::tibble()}}
}
\details{
The \code{list} output format ensures that numeric parameters are treated as is. On
the contrary, the \code{\link[tibble:tibble]{tibble::tibble()}} output will return all parameters as characters if
at least one is found to be character. This behavior is used to ensure a consistant expected
output.
The parameter names are partially matched, so the user can retreive their values
without the need of perfectly knowing their names.
The configuration file is named "ArchimedConfiguration.properties".
}
\examples{
 \dontrun{
 # Read two parameters, with partial matching (latit instead of latitude):
 read_config(file = "Archimed_July_2018/app_parameters/ArchimedConfiguration.properties",
             parameter = c("latit","altitude"))
 # read all parameters:
 read_config(file = "Archimed_July_2018/app_parameters/ArchimedConfiguration.properties")

 }

}
