% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_files.R
\name{read_ops}
\alias{read_ops}
\title{read ops file}
\usage{
read_ops(file)
}
\arguments{
\item{file}{the path to the ops file}
}
\value{
A list of two \code{data.frame}:
\itemize{
\item dimensions: the plot dimensions (xmin, ymin, zmin, xmax, ymax);
\item plants= the informations about the plants positions, path to the opf, ...
}
}
\description{
read ops file
}
\examples{
\dontrun{
read_ops("path_to_the_ops/dummy.ops")
}
}
