% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_outputs.R
\name{read_out_node}
\alias{read_out_node}
\title{Read node-scale output}
\usage{
read_out_node(path, duration = NULL)
}
\arguments{
\item{path}{File path}

\item{duration}{data.frame of two columns: the time-step, and its duration in seconds}
}
\value{
The ARCHIMED outputs for all nodes
}
\description{
Read the ARCHIMED output at node-scale
}
\details{
the duration data.frame must have two columns: step (step index) and timestep
(duration in seconds)
}
\examples{
\dontrun{
read_out_node(path= "output/nodes_values.csv")
}

}
