% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_outputs.R
\name{read_out_plot}
\alias{read_out_plot}
\title{Read plot-scale output}
\usage{
read_out_plot(path, sheet = "Mir")
}
\arguments{
\item{path}{File path}

\item{sheet}{Desired Excel sheet}
}
\value{
The plot-scale ARCHIMED outputs
}
\description{
Read the ARCHIMED output aggregated at plot-scale
}
\examples{
\dontrun{
read_out_plot(path= "output/2018.11.15_14-18-11_mirdata_46.xlsx", sheet= "Mir")
}

}
