% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_voxels.R
\name{read_vox}
\alias{read_vox}
\title{Read voxel file}
\usage{
read_vox(input)
}
\arguments{
\item{input}{File path (including file name). Can also be anything compatible with
\code{data.table::fread()}.}
}
\value{
The voxel data
}
\description{
Read the voxel file, either input or output from ARCHIMED
}
\details{
This function uses \code{read_voxel_space()} to find the voxel space and set it as an
attribute: \code{voxel_space}.
}
\examples{
\dontrun{
# Read a voxel file:
vox= read_vox(path= "output/voxelized_scene.vox")
# Find the voxel space:
attr(vox, "voxel_space")
}

}
