% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_config}
\alias{set_config}
\title{Set the configuration}
\usage{
set_config(file, parameter, value)
}
\arguments{
\item{file}{The path to the ARCHIMED configuration file}

\item{parameter}{Parameter name}

\item{value}{The new value of the parameter}
}
\value{
Nothing. Writes in the file directly.
}
\description{
Set a parameter value for the ARCHIMED configuration file
}
\details{
The configuration file is named "ArchimedConfiguration.properties".
}
