% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{temperature_dynamic}
\alias{temperature_dynamic}
\title{Temperature dynamic}
\usage{
temperature_dynamic(Tmax, Tmin, per, shift = 3 * pi/2)
}
\arguments{
\item{Tmax}{Maximum air temperature}

\item{Tmin}{Minimum air temperature}

\item{per}{The sine curve periodicity}

\item{shift}{Phase shift, or when does Tmin/Tmax does occur ?}
}
\value{
A simulated temperature dynamics over a period of time
}
\description{
Simulate a temperature dynamic over a time period
}
\examples{
# Temperature dynamic over a day:
temperature_dynamic(30,18,24,3*pi/2)
}
