% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_voxels.R
\name{voxel_space}
\alias{voxel_space}
\title{Voxel space}
\usage{
voxel_space(
  coord_min = c(0, 0, 0),
  voxel_size = NULL,
  n_voxel = NULL,
  coord_max = NULL
)
}
\arguments{
\item{coord_min}{Vector of the minimum 3D coordinates of the voxel space (point of origin, see details)}

\item{voxel_size}{Vector of the lengths of the voxel side in x, y and z.}

\item{n_voxel}{The number of voxels in c, y and z (optional, see details)}

\item{coord_max}{Vector of the maximum 3D coordinates of the voxel space}
}
\value{
A voxel space
}
\description{
Defines the voxel space using minimum and maximum coordinates
}
\details{
The 3D coordinates have the form of c(x,y,z). The user chooses to compute the voxel space from
a pair of arguments: either voxel_size and n_voxel, either voxel_size and coord_max, or n_voxel and
coord_max.
}
\examples{
voxel_space(voxel_size= c(1,1,1), n_voxel= c(10,10,20))
}
