% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_outputs.R
\name{read_out_node}
\alias{read_out_node}
\title{Read node-scale output}
\usage{
read_out_node(path)
}
\arguments{
\item{path}{File path}
}
\value{
The ARCHIMED outputs for all nodes
}
\description{
Read the ARCHIMED output at node-scale
}
\examples{
\dontrun{
read_out_node(path= "output/nodes_values.csv")
}

}
