% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_outputs.R
\name{read_vox}
\alias{read_vox}
\title{Read voxel file}
\usage{
read_vox(input)
}
\arguments{
\item{input}{File path (including file name). Can also be anything compatible with
\code{data.table::fread()}.}
}
\value{
The voxel data
}
\description{
Read the voxel file, either input or output from ARCHIMED
}
\examples{
\dontrun{
read_vox(path= "output/voxelized_scene.vox")
}

}
