% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_archimed.R
\name{run_archimed}
\alias{run_archimed}
\title{Run ARCHIMED}
\usage{
run_archimed(path = getwd(), memory = 4096)
}
\arguments{
\item{path}{Path to the ARCHIMED root}

\item{memory}{The memory allocated to the JVM (Java Virtual Machine)}
}
\value{
The function prints the model outputs to the console and returns
\code{TRUE} if ARCHIMED ran successfully, or an error code if any problem occured
}
\description{
Run ARCHIMED
}
\examples{
\dontrun{
run_archimed(path= "Archimed")
}
}
