% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_archimed.R
\name{run_archimed}
\alias{run_archimed}
\title{Run ARCHIMED}
\usage{
run_archimed(exe = file.path(getwd(), "archimed-lib_florian-1.0.0.jar"),
  memory = 4096)
}
\arguments{
\item{exe}{Path to the ARCHIMED executable file including its name}

\item{memory}{The memory allocated to the JVM (Java Virtual Machine)}
}
\value{
The function prints the model outputs to the console and returns
\code{TRUE} if ARCHIMED ran successfully, or an error code if any problem occured
}
\description{
Run ARCHIMED
}
\examples{
\dontrun{
run_archimed(path= "Archimed/archimed.jar")
}
}
